# WSGmed/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "API for managing medical app.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Medical API Server"
        }
    ],
    "paths": {
...
        "/api/medications": {
            "get": {
                "tags": [
                    "Medications"
                ],
                "summary": "Get patient's medications for today",
                "description": "Returns a list of medications assigned to the logged-in patient for today based on start_date/end_date and marks them as taken if a confirmation exists for today.",
                "operationId": "getMedicationsByDate",
                "responses": {
                    "200": {
                        "description": "List of medications",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Medications retrieved successfully."
                                        },
                                        "data": {
                                            "properties": {
                                                "med_taken": {
                                                    "type": "integer",
                                                    "example": 2
                                                },
                                                "med_all": {
                                                    "type": "integer",
                                                    "example": 3
                                                },
                                                "medications": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "name": {
                                                                "type": "string",
                                                                "example": "Aspirin"
                                                            },
                                                            "info": {
                                                                "type": "string",
                                                                "example": "Pain reliever and fever reducer."
                                                            },
                                                            "patient_medication_id": {
                                                                "type": "integer",
                                                                "example": 10
                                                            },
                                                            "dosage": {
                                                                "type": "string",
                                                                "example": "1 tabletka"
                                                            },
                                                            "is_taken": {
                                                                "type": "boolean",
                                                                "example": false
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized access",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        },
                                        "code": {
                                            "type": "integer",
                                            "example": 10002
                                        }
                                    },
                                    "type": "object"
                                },
                                "example": {
                                    "success": false,
                                    "message": "Authentication token not provided.",
                                    "code": 10002
                                }
                            }
                        }
                    },
                    "503": {
                        "description": "Service Unavailable - Database connection issues",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "The service is temporarily unavailable. Please try again later."
                                        },
                                        "code": {
                                            "type": "integer",
                                            "example": 19002
                                        }
                                    },
                                    "type": "object"
                                },
                                "example": {
                                    "success": false,
                                    "message": "The service is temporarily unavailable. Please try again later.",
                                    "code": 19002
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "An unexpected error occurred on the server."
                                        },
                                        "code": {
                                            "type": "integer",
                                            "example": 19001
                                        }
                                    },
                                    "type": "object"
                                },
                                "example": {
                                    "success": false,
                                    "message": "An unexpected error occurred on the server.",
                                    "code": 19001
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}